% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSSA.R
\docType{package}
\name{POSSA-package}
\alias{POSSA}
\alias{POSSA-package}
\title{POSSA: Power Simulation for Sequential Analysis}
\description{
Power simulation for sequential analysis. Blah.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gasparl/possa}
}

}
\author{
\strong{Maintainer}: Gáspár Lukács \email{lkcsgaspar@gmail.com}

}
\keyword{internal}
