% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Power calculation}
\usage{
pow(
  p_values,
  alpha_locals = NULL,
  alpha_global = 0.05,
  fut_locals = NULL,
  adjust = NULL,
  adj_init = NULL,
  group_by = NULL,
  alpha_locals_extra = NULL,
  design_fix = NULL,
  design_seq = TRUE,
  descr_cols = TRUE,
  descr_func = summary,
  round_to = 3,
  multi_logic = "all",
  multi_logic_fut = "all",
  staircase_steps = NULL,
  alpha_precision = 5,
  seed = 8,
  hush = FALSE
)
}
\arguments{
\item{p_values}{Data frame.}

\item{round_to}{Number \code{\link[=ro]{to round}} to.}

\item{hush}{Logical. If \code{TRUE}, prevents printing any details to console.}
}
\value{
Returns \code{\link{data.frame}}, etc.
}
\description{
Calculates power (and local alphas) based on given global alpha
 and simulated p values.
}
\details{

}
\note{
Some notes.
}
\examples{
# some pow

}
\references{
Lakens, D. (2014). Performing high-powered studies efficiently with sequential
analyses: Sequential analyses. European Journal of Social Psychology, 44(7),
701–710. \doi{https://doi.org/10.1002/ejsp.2023}
}
\seealso{
\code{\link{sim}}
}
