% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.possa_pow_df.R
\name{print.possa_pow_df}
\alias{print.possa_pow_df}
\title{Print pow results data frame}
\usage{
\method{print}{possa_pow_df}(df_pow, round_to = NA, possa_title = TRUE, ...)
}
\arguments{
\item{round_to}{Number of fractional digits to round to, for the displayed
numbers. The default is the value passed from the
\code{\link[POSSA:pow]{POSSA::pow}} function (stored in the returned list).}

\item{pow_list}{The \code{\link{list}} returned by the
\code{\link[POSSA:pow]{POSSA::pow}} function.}
}
\description{
Prints, in a readable manner, the main information from any of
 the data frames containing power information from the list created by the
 \code{\link[POSSA:pow]{POSSA::pow}} function. This is an extension (method)
 of the base R \code{\link[print]} function, so it can be called simply as
 \code{print()}.
}
\seealso{
\code{\link{pow}}, \code{\link{print.possa_pow_list}}
}
