% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.possa_pow_df.R
\name{print.possa_pow_list}
\alias{print.possa_pow_list}
\title{Print pow results}
\usage{
\method{print}{possa_pow_list}(x, round_to = NA, ...)
}
\arguments{
\item{x}{The \code{\link{list}} returned by the
\code{\link[POSSA:pow]{POSSA::pow}} function.}

\item{round_to}{Number of fractional digits to round to, for the displayed
numbers. The default is the value passed from the
\code{\link[POSSA:pow]{POSSA::pow}} function (stored in the returned list).}

\item{...}{(Allow additional arguments for technical reasons.)}
}
\value{
Returns nothing (\code{NULL}); this method serves only to print
 information to the console.
}
\description{
Prints, in a readable manner, the main information from the list
 created by the \code{\link[POSSA:pow]{POSSA::pow}} function, calling
 \code{\link{print.possa_pow_df}} for each of the POSSA power information
 data frames in the list. This is an extension (method) of the base R
 \code{\link{print}} function, so it can be called simply as \code{print()}.
}
\seealso{
\code{\link{pow}}, \code{\link{print.possa_pow_df}}
}
