% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POSSA.R
\docType{package}
\name{POSSA-package}
\alias{POSSA}
\alias{POSSA-package}
\title{POSSA: Power Simulation for Sequential Analysis}
\description{
This R package serves to calculate, via simulation, power and appropriate
"stopping" alpha boundaries (and/or futility bounds) for sequential analyses
(i.e., "group sequential design") as well as for multiple hypotheses
(multiple tests included in an analysis), given any specified global error
rate. This enables the sequential use of practically any significance test,
as long as the underlying data can be simulated in advance to a reasonable
approximation. Note: Since the specific functions are rather complex, it is
much easier to understand the procedure with the examples given in the
vignettes (see via https://github.com/gasparl/possa).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gasparl/possa}
}

}
\author{
\strong{Maintainer}: Gáspár Lukács \email{lkcsgaspar@gmail.com}

}
\keyword{internal}
