% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulation procedure}
\usage{
sim(f_sample, n_obs, f_test, n_iter = 1000, seed = 8)
}
\arguments{
\item{f_sample}{Function.}

\item{n_iter}{Number of iterations.}
}
\value{
Returns \code{\link{data.frame}}, etc.
}
\description{
Simulation procedure, running the tests in many iterations to get
 p values.
}
\details{

}
\note{
Some notes.
}
\examples{
# some sim

}
\references{
Lakens, D. (2014). Performing high-powered studies efficiently with sequential
analyses: Sequential analyses. European Journal of Social Psychology, 44(7),
701–710. \doi{https://doi.org/10.1002/ejsp.2023}
}
\seealso{
\code{\link{pow}}
}
